Follow langpack spec changes: version -> revision (bug 1134147)
"version" was ambiguous and incoherent with the "version" field
at the top-level of the manifest, since FxOS expected it to be an
int, in order to compare multiple langpacks providing the same
languages.

It was renamed to "revision" in bug 1133725 so we follow that and
change the expected type to int.